Week2/12_factor_app/requirements.backend.txt
fastapi==0.115.0 
uvicorn==0.30.6
pydantic==2.9.2 
torch==2.5.0 
torchvision==0.20.0 
pandas==2.2.3 
pillow==10.4.0
numpy==2.1.1 
pytest==8.3.3 
httpx==0.27.2
loguru